Fix state in cards
