+++ frontend/frontend/src/components/MainPage/mainpage.jsx
@@ -9,6 +9,7 @@ import { Lang } from './Lang/Lang';
 import { UserId } from '../UserId/UserId';
 
 import { t } from 'ttag';
+
 //Font Awesome
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faPaperPlane } from "@fortawesome/free-solid-svg-icons";
