+++ frontend/frontend/src/components/Submenu/Game/game.css
@@ -1,5 +1,6 @@
 .site-main-game {
     height: 100%;
+    min-height: 100vh;
     background: linear-gradient(103.9deg, rgba(0, 0, 0, 0.56) -3.06%, rgba(194, 194, 194, 0) 103.09%);
     padding: 20px 0;
 }
